Move retransmits to mostly driven by a retransmit loop
This re-complexs a lot of things. But also allows for
nice things like retransmitting "dead" frames in-order.

Since the receiveQueue on the other side only has 4096 slots
if you don't transmit them in order, then you may have
to send them again because the frames got dropped due
to being too far in the future.