Overhaul of the sending loop.
This new sending loop machineary avoids blocking. This allows
connections to lock up (or just hit the back of it's TCP SendQ)
without the rest of the application locking up with it.

This also populates data needed for the retransmission loop
to avoid retransmissions from being send down the same subflow
over and over again. Allowing very rapid recovery from a subflow
failing.